requirements.txt
dash
dash-bootstrap-components
diskcache
multiprocess
psutil
pandas
python-dotenv
gunicorn
supabase
fredapi